+++ contracts/croncat-tasks/src/tests/task_tests.rs
@@ -1030,8 +1030,8 @@ fn remove_tasks_with_queries_success() {
             cw20: None,
             coin: [Some(coin(5, DENOM)), None],
             gas: 50_000,
-            agent_fee: u64::default(),
-            treasury_fee: u64::default(),
+            agent_fee: u16::default(),
+            treasury_fee: u16::default(),
             gas_price: GasPrice::default(),
         },
     };
@@ -1058,8 +1058,8 @@ fn remove_tasks_with_queries_success() {
             cw20: None,
             coin: [Some(coin(5, DENOM)), None],
             gas: 50_000,
-            agent_fee: u64::default(),
-            treasury_fee: u64::default(),
+            agent_fee: u16::default(),
+            treasury_fee: u16::default(),
             gas_price: GasPrice::default(),
         },
     };
